requirements.txt
openai>=1.50.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
fastapi>=0.100.0
uvicorn>=0.23.0
chromadb>=0.4.20
langchain-openai>=0.1.0
langgraph>=0.0.26